+++ config/config.go
@@ -256,18 +256,18 @@ func Configure(args []string, readFunc FileReadFunc) (MainConfig, map[string]str
 		}
 	}
 
-	if err := sanitizeValues("destination attribute", cfg.Destination.Attributes); err != nil {
+	if err := sanitizeValues("destination attribute", false, cfg.Destination.Attributes); err != nil {
 		return MainConfig{}, nil, nil, err
 	}
-	if err := sanitizeValues("destination header", cfg.Destination.Headers); err != nil {
+	if err := sanitizeValues("destination header", true, cfg.Destination.Headers); err != nil {
 		return MainConfig{}, nil, nil, err
 	}
 
 	if cfg.Diagnostics.Endpoint != "" {
-		if err := sanitizeValues("diagnostics attribute", cfg.Diagnostics.Attributes); err != nil {
+		if err := sanitizeValues("diagnostics attribute", false, cfg.Diagnostics.Attributes); err != nil {
 			return MainConfig{}, nil, nil, err
 		}
-		if err := sanitizeValues("diagnostics header", cfg.Diagnostics.Headers); err != nil {
+		if err := sanitizeValues("diagnostics header", true, cfg.Diagnostics.Headers); err != nil {
 			return MainConfig{}, nil, nil, err
 		}
 	}
@@ -289,9 +289,12 @@ func sanitize(val string) string {
 	return val
 }
 
-func sanitizeValues(kind string, values map[string]string) error {
+func sanitizeValues(kind string, downcaseKeys bool, values map[string]string) error {
 	for origKey, value := range values {
 		key := sanitize(origKey)
+		if downcaseKeys {
+			key = strings.ToLower(key)
+		}
 		if key == "" {
 			return fmt.Errorf("empty %s key", kind)
 		}